+++ src/modules/account/components/MaterialRequestDetailBlade/hooks/useGetMaterialRequestConversationInfinite.ts
@@ -1,5 +1,5 @@
-import type { MaterialRequestMessage } from '@account/components/MaterialRequestDetailBlade/MaterialRequestConversation.types';
 import { MaterialRequestsService } from '@material-requests/services';
+import type { MaterialRequestMessage } from '@material-requests/types';
 import { QUERY_KEYS } from '@shared/const/query-keys';
 import type { IPagination } from '@studiohyperdrive/pagination';
 import {
