+++ src/modules/account/components/MaterialRequestDetailBlade/hooks/useSendMaterialRequestMessage.ts
@@ -1,5 +1,5 @@
-import type { MaterialRequestMessage } from '@account/components/MaterialRequestDetailBlade/MaterialRequestConversation.types';
 import { MaterialRequestsService } from '@material-requests/services';
+import type { MaterialRequestMessage } from '@material-requests/types';
 import { QUERY_KEYS } from '@shared/const/query-keys';
 import { type UseMutationResult, useMutation, useQueryClient } from '@tanstack/react-query';
