+++ src/styles/components/_index.scss
@@ -1,7 +1,6 @@
 @use "alert" as *;
 @use "avatar" as *;
 @use "badge" as *;
-@use "blade" as *;
 @use "box" as *;
 @use "breadcrumbs" as *;
 @use "button" as *;
